+++ processor/src/main/kotlin/io/mcarle/konvert/processor/konvert/KonvertCodeGenerator.kt
@@ -7,7 +7,6 @@ import com.squareup.kotlinpoet.FunSpec
 import com.squareup.kotlinpoet.KModifier
 import com.squareup.kotlinpoet.TypeSpec
 import com.squareup.kotlinpoet.ksp.toTypeName
-import io.mcarle.konvert.converter.api.classDeclaration
 import io.mcarle.konvert.processor.codegen.CodeBuilder
 import io.mcarle.konvert.processor.codegen.CodeGenerator
 import io.mcarle.konvert.processor.validated
@@ -29,16 +28,32 @@ object KonvertCodeGenerator {
             converter.mapKSClassDeclaration.simpleName.asString(),
         )
 
+        if (converter.sourceTypeReference.toString() != converter.sourceType.toString()) {
+            // add import alias
+            codeBuilder.addImport(converter.sourceType, converter.sourceTypeReference.toString())
+        }
+        val targetClassImportName = if (converter.targetTypeReference.toString() != converter.targetType.toString()) {
+            // add import alias
+            val alias = converter.targetTypeReference.toString()
+            codeBuilder.addImport(converter.targetType, alias)
+            alias
+        } else if (converter.sourceTypeReference.toString() == converter.targetTypeReference.toString()) {
+            null
+        } else {
+            converter.targetClassDeclaration.simpleName.asString()
+        }
+
         codeBuilder.addFunction(
             funSpec = FunSpec.builder(converter.mapFunctionName)
                 .addModifiers(KModifier.OVERRIDE)
-                .returns(converter.targetType.toTypeName())
-                .addParameter(converter.paramName, converter.sourceType.toTypeName())
+                .returns(converter.targetTypeReference.toTypeName())
+                .addParameter(converter.paramName, converter.sourceTypeReference.toTypeName())
                 .addCode(
                     mapper.generateCode(
                         converter.annotation.mappings.asIterable().validated(converter.mapKSFunctionDeclaration, logger),
                         converter.annotation.constructor,
                         converter.paramName,
+                        targetClassImportName,
                         converter.sourceClassDeclaration,
                         converter.targetClassDeclaration,
                         converter.mapKSFunctionDeclaration